Cambio los tipos a float

+++ division.py
@@ -1,4 +1,4 @@
-def division(a: int, b: int):
+def division(a: float, b: float):
   if b == 0:
     raise ValueError("No se puede dividir por cero.")
